ai-service/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20
python-dotenv==1.1.0
langchain==0.3.25
langchain-community==0.3.24
langchain-chroma==0.2.4
langchain-groq==0.3.2
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.8
langchain-unstructured==0.1.6
langgraph==0.4.5
sentence-transformers==4.1.0
chromadb==1.0.10
unstructured[all-docs]==0.17.2
